+++ client/src/services/api.js
@@ -137,7 +137,7 @@ export async function deleteSlideshow (params) {
   return QL.remove('slideshow', params, ['result']);
 }
 export async function queryOrder (params) {
-  return QL.find_many('order', deleteEmptyProperty(params), [  'orderPay_id',  'productTotalPay',  'count',  'id',  'orderStatus',  'createdAt',  'orderTotalPay',  'orderLogistics_id',  'updatedAt',  'remark',  'orderlogistic.deliveryTime',  'orderlogistic.serviceStore',  'orderlogistic.updatedAt',  'orderlogistic.logisticsFee',  'orderlogistic.expressId',  'orderlogistic.createdAt',  'orderlogistic.consigneeTel',  'orderlogistic.id',  'orderlogistic.expressName',  'orderlogistic.consignAddress',  'orderlogistic.LogisticsStatus',  'orderlogistic.consigneeName',  'orderproduct.updatedAt',  'orderproduct.productColor',  'orderproduct.unit',  'orderproduct.productSize',  'orderproduct.orderPay',  'orderproduct.createdAt',  'orderproduct.productImg',  'orderproduct.productName',  'orderproduct.productPrice',  'orderproduct.id',  'orderproduct.count',  'orderproduct.productPay',  'orderproduct.orderPay_id',  'userAddress_id.address',  'userAddress_id.updatedAt',  'userAddress_id.telephone',  'userAddress_id.default',  'userAddress_id.city',  'userAddress_id.username',  'userAddress_id.postcode',  'userAddress_id.createdAt',  'userAddress_id.deletedAt',  'userAddress_id.id',  'userAddress_id.area',  'userAddress_id.province',  'user.email',  'user.updatedAt',  'user.password',  'user.telephone',  'user.username',  'user.createdAt',  'user.id',  'user.userData_id',  ]);
+  return QL.find_many('order', deleteEmptyProperty(params), [  'orderPay_id',  'productTotalPay',  'count',  'id',  'orderStatus',  'createdAt',  'orderTotalPay',  'orderLogistics_id',  'updatedAt',  'remark',  'user_id.id']);
 }
 export async function queryOneOrder (params) {
   return QL.find_one('order', params, [  'orderPay_id',  'productTotalPay',  'count',  'id',  'orderStatus',  'createdAt',  'orderTotalPay',  'orderLogistics_id',  'updatedAt',  'remark',  'orderlogistic.deliveryTime',  'orderlogistic.serviceStore',  'orderlogistic.updatedAt',  'orderlogistic.logisticsFee',  'orderlogistic.expressId',  'orderlogistic.createdAt',  'orderlogistic.consigneeTel',  'orderlogistic.id',  'orderlogistic.expressName',  'orderlogistic.consignAddress',  'orderlogistic.LogisticsStatus',  'orderlogistic.consigneeName',  'orderproduct.updatedAt',  'orderproduct.productColor',  'orderproduct.unit',  'orderproduct.productSize',  'orderproduct.orderPay',  'orderproduct.createdAt',  'orderproduct.productImg',  'orderproduct.productName',  'orderproduct.productPrice',  'orderproduct.id',  'orderproduct.count',  'orderproduct.productPay',  'orderproduct.orderPay_id',  'userAddress_id.address',  'userAddress_id.updatedAt',  'userAddress_id.telephone',  'userAddress_id.default',  'userAddress_id.city',  'userAddress_id.username',  'userAddress_id.postcode',  'userAddress_id.createdAt',  'userAddress_id.deletedAt',  'userAddress_id.id',  'userAddress_id.area',  'userAddress_id.province',  'user.email',  'user.updatedAt',  'user.password',  'user.telephone',  'user.username',  'user.createdAt',  'user.id',  'user.userData_id',  ]);
